requirements.txt
numpy
paho-mqtt
pyproj
pyzmq
Flask